# psd-render-api/openapi-current.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PSD 渲染服务平台 API",
    "description": "## 概述\nPSD 模板渲染服务平台对外提供 PSD 模板管理、资产上传、渲染任务提交与查询的 HTTP API。\n\n## 鉴权\n所有 `/v1/*` 接口需在 `Authorization` 头携带 API Key（格式：`Bearer sk_live_xxx`）。\nAPI Key 通过 Admin 控制台创建，每个 Key 可配置：\n- 限流（rateLimitPerMin）：每分钟最大请求数\n- 日配额（quotaPerDay）：每天最大请求数\n- 作用域（scopes）：限定可调用的接口\n- IP 白名单（ipWhitelist）\n- 默认 Webhook URL / Secret\n\n## 幂等性\n`POST /v1/render-jobs` 支持通过 `Idempotency-Key` 头实现幂等：\n同租户同 key 重复提交将返回相同任务，不重复创建。\n\n## Webhook\n任务终态（succeeded / failed / cancelled）时自动投递 Webhook 到提交时声明的 `webhookUrl`。\n投递包含 HMAC-SHA256 签名头 `X-Render-Signature`，可校验来源。\n失败自动重试，指数退避：1m / 5m / 15m / 60m / 240m / 960m。\n\n## 错误响应\n所有错误响应统一为 `{ \"error\": \"<CODE>\", \"message\": \"<描述>\", \"details\"?: [...] }`。",
    "version": "0.3.0",
    "contact": {
      "name": "PSD 渲染服务平台"
    }
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API Key（格式：sk_live_xxx）",
        "bearerFormat": "sk_live_xxx"
      },
      "adminSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Admin Session Token（登录后获得）",
        "bearerFormat": "session_token"
      },
      "workerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Worker 访问令牌（注册后获得）",
        "bearerFormat": "wkr_xxx"
      },
      "storageSignedToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "存储签名令牌（格式：Bearer <token>，由 generateUploadUrl/generateDownloadUrl 签发，仅 local 模式可用。也支持 ?token= 查询参数，但不推荐）"
      },
      "workerRegisterSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Worker 注册凭据（格式：Bearer <WORKER_REGISTER_SECRET> 或一次性配对码/token，也可通过 X-Worker-Register-Secret 头传递）"
      }
    },
    "schemas": {
      "def-0": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误描述"
          },
          "details": {
            "type": "array",
            "description": "可选，详细错误信息"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse"
      }
    }
  },
  "paths": {
...
    "/v1/assets/upload-url": {
      "post": {
        "summary": "获取资产上传地址",
        "tags": [
          "assets"
        ],
        "description": "获取输入文件预签名上传地址与 assetId。仅支持 JPG/PNG/JPEG，累计 ≤150MB，输入保留 1 天。",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fileName"
                ],
                "properties": {
                  "fileName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255,
                    "pattern": "^[a-zA-Z0-9._\\-]+$",
                    "description": "原始文件名（含扩展名，仅允许字母/数字/点/下划线/连字符，≤255 字符）"
                  },
                  "mimeType": {
                    "type": "string",
                    "maxLength": 128,
                    "description": "可选，未提供则按扩展名推断（image/jpeg / image/png），≤128 字符"
                  },
                  "sizeBytes": {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 1,
                    "maximum": 157286400,
                    "description": "可选，文件大小（字节，上限 150MB）"
                  }
                }
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "assetId",
                    "uploadUrl",
                    "method",
                    "headers",
                    "objectKey",
                    "expiresAt"
                  ],
                  "properties": {
                    "assetId": {
                      "type": "string",
                      "description": "资产 ID（art_xxx），后续提交任务时使用"
                    },
                    "uploadUrl": {
                      "type": "string",
                      "description": "预签名 PUT 上传地址（有效期 10 分钟）"
                    },
                    "method": {
                      "type": "string",
                      "description": "上传方法（PUT）"
                    },
                    "headers": {
                      "type": "object",
                      "description": "上传时必须携带的请求头（如 Content-Type）"
                    },
                    "objectKey": {
                      "type": "string",
                      "description": "存储对象 key"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "上传地址过期时间"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "API Key 无效或缺失"
                }
              }
            }
          },
          "403": {
            "description": "IP 白名单拒绝 / 作用域不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "IP 白名单拒绝 / 作用域不足"
                }
              }
            }
          },
          "429": {
            "description": "触发限流（rateLimitPerMin）或日配额耗尽（quotaPerDay）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "触发限流（rateLimitPerMin）或日配额耗尽（quotaPerDay）"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}